requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
apache-tvm-ffi==0.1.8.post2
astor==0.8.1
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.8
blinker==1.9.0
boto3==1.42.54
botocore==1.42.54
cachetools==7.0.1
cbor2==5.8.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
cloudpickle==3.1.2
compressed-tensors==0.13.0
cryptography==46.0.5
cuda-bindings==13.1.1
cuda-pathfinder==1.3.4
cuda-python==13.1.1
cupy-cuda12x==14.0.1
depyf==0.20.0
dill==0.4.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
einops==0.8.2
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.129.2
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
filelock==3.18.0
flashinfer-python==0.6.1
Flask==3.1.0
frozenlist==1.6.0
fsspec==2025.3.2
gguf==0.17.1
grpcio==1.78.1
grpcio-reflection==1.78.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.10
ijson==3.4.0.post0
imageio==2.37.0
interegular==0.3.3
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
jmespath==1.1.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.2.2
lazy_loader==0.4
llguidance==1.3.0
llvmlite==0.44.0
lm-format-enforcer==0.11.3
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mcp==1.26.0
mdurl==0.1.2
mistral_common==1.9.1
model-hosting-container-standards==0.1.13
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.20.0
multidict==6.4.3
networkx==3.4.2
ninja==1.13.0
numba==0.61.2
numpy==2.2.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cudnn-frontend==1.18.0
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-cutlass-dsl==4.4.0
nvidia-cutlass-dsl-libs-base==4.4.0
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.21.0
openai-harmony==0.0.8
opencv-python-headless==4.13.0.92
outlines_core==0.2.11
packaging==25.0
partial-json-parser==0.2.1.1.post7
pillow==11.2.1
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.24.1
propcache==0.3.1
protobuf==6.33.5
psutil==7.0.0
py-cpuinfo==9.0.0
pybase64==1.4.3
pycountry==26.2.16
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pynvml==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==4.0.0
python-multipart==0.0.22
PyYAML==6.0.2
pyzmq==27.1.0
ray==2.54.0
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
rich==14.3.3
rich-toolkit==0.19.4
rignore==0.7.6
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.2
sentence-transformers==5.2.3
sentencepiece==0.2.1
sentry-sdk==2.53.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==3.2.0
starlette==0.52.1
supervisor==4.3.0
sympy==1.13.3
tabulate==0.9.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.6
triton==3.5.1
twilio==9.5.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.4.0
uvicorn==0.41.0
uvloop==0.22.1
vllm==0.15.1
watchfiles==1.1.1
websockets==16.0
Werkzeug==3.1.3
xgrammar==0.1.29
yarl==1.20.0
